# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Job {
  id: number;
  name: string;
  category: string;
  available: boolean;
  /** @nullable */
  discordLink?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface JobInput {
  name: string;
  category: string;
  available?: boolean;
  /** @nullable */
  discordLink?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface JobUpdate {
  name?: string;
  category?: string;
  available?: boolean;
  /** @nullable */
  discordLink?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface IllegalOrg {
  id: number;
  name: string;
  status: string;
  activities: string[];
  /** @nullable */
  discordLink?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  segment: string;
}

export interface IllegalOrgInput {
  name: string;
  status: string;
  activities: string[];
  /** @nullable */
  discordLink?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  segment?: string;
}

export interface IllegalOrgUpdate {
  name?: string;
  status?: string;
  activities?: string[];
  /** @nullable */
  discordLink?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  segment?: string;
}

export interface LoreEntry {
  id: number;
  title: string;
  summary: string;
  sortOrder: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface LoreInput {
  title: string;
  summary: string;
  sortOrder?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface LoreUpdate {
  title?: string;
  summary?: string;
  sortOrder?: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface PatchNote {
  id: number;
  version: string;
  date: string;
  category: string;
  changes: string[];
  sortOrder: number;
}

export interface PatchNoteInput {
  version: string;
  date: string;
  category: string;
  changes: string[];
  sortOrder?: number;
}

export interface PatchNoteUpdate {
  version?: string;
  date?: string;
  category?: string;
  changes?: string[];
  sortOrder?: number;
}

export interface GalleryItem {
  id: number;
  imageUrl: string;
  description: string;
  sortOrder: number;
}

export interface GalleryItemInput {
  imageUrl: string;
  description: string;
  sortOrder?: number;
}

export interface GalleryItemUpdate {
  imageUrl?: string;
  description?: string;
  sortOrder?: number;
}

export interface Staff {
  id: number;
  pseudo: string;
  role: string;
  description: string;
  category: string;
  /** @nullable */
  avatarUrl?: string | null;
  sortOrder: number;
}

export interface StaffInput {
  pseudo: string;
  role: string;
  description: string;
  category: string;
  /** @nullable */
  avatarUrl?: string | null;
  sortOrder?: number;
}

export interface StaffUpdate {
  pseudo?: string;
  role?: string;
  description?: string;
  category?: string;
  /** @nullable */
  avatarUrl?: string | null;
  sortOrder?: number;
}

export interface AdminLoginInput {
  username: string;
  password?: string;
}

export interface AdminMe {
  loggedIn: boolean;
}