# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/gallery:
    post:
      operationId: createGalleryItem
      tags: [admin]
      summary: Create a gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemInput"
      responses:
        "201":
          description: Created gallery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
        "401":
          description: Unauthorized
  /admin/gallery/{id}:
    patch:
      operationId: updateGalleryItem
      tags: [admin]
      summary: Update a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemUpdate"
      responses:
        "200":
          description: Updated gallery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
        "401":
          description: Unauthorized
        "404":
          description: Gallery item not found
    delete:
      operationId: deleteGalleryItem
      tags: [admin]
      summary: Delete a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Gallery item not found
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
        "401":
          description: Unauthorized
...
  /admin/jobs:
    post:
      operationId: createJob
      tags: [admin]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
  /admin/jobs/{id}:
    patch:
      operationId: updateJob
      tags: [admin]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
        "404":
          description: Job not found
    delete:
      operationId: deleteJob
      tags: [admin]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Job not found
  /admin/illegal:
    post:
      operationId: createIllegalOrg
      tags: [admin]
      summary: Create an illegal organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IllegalOrgInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IllegalOrg"
        "401":
          description: Unauthorized
  /admin/illegal/{id}:
    patch:
      operationId: updateIllegalOrg
      tags: [admin]
      summary: Update an illegal organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IllegalOrgUpdate"
      responses:
        "200":
          description: Updated illegal organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IllegalOrg"
        "401":
          description: Unauthorized
        "404":
          description: Illegal organization not found
    delete:
      operationId: deleteIllegalOrg
      tags: [admin]
      summary: Delete an illegal organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Illegal organization not found
  /admin/lore:
    post:
      operationId: createLore
      tags: [admin]
      summary: Create a lore entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoreInput"
      responses:
        "201":
          description: Created lore entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoreEntry"
        "401":
          description: Unauthorized
  /admin/lore/{id}:
    patch:
      operationId: updateLore
      tags: [admin]
      summary: Update a lore entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoreUpdate"
      responses:
        "200":
          description: Updated lore entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoreEntry"
        "401":
          description: Unauthorized
        "404":
          description: Lore entry not found
    delete:
      operationId: deleteLore
      tags: [admin]
      summary: Delete a lore entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Lore entry not found
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, name, category, available]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        available: { type: boolean }
        discordLink: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
    JobInput:
      type: object
      required: [name, category]
      properties:
        name: { type: string }
        category: { type: string }
        available: { type: boolean }
        discordLink: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
    JobUpdate:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        available: { type: boolean }
        discordLink: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
    IllegalOrg:
      type: object
      required: [id, name, status, activities, segment]
      properties:
        id: { type: integer }
        name: { type: string }
        status: { type: string }
        activities:
          type: array
          items: { type: string }
        discordLink: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        segment:
          type: string
          description: Regroupement affiché sur le site (Gang, Organisation, Indépendant).
    IllegalOrgInput:
      type: object
      required: [name, status, activities]
      properties:
        name: { type: string }
        status: { type: string }
        activities:
          type: array
          items: { type: string }
        discordLink: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        segment:
          type: string
          description: Gang | Organisation | Indépendant (défaut Organisation si omis).
    IllegalOrgUpdate:
      type: object
      properties:
        name: { type: string }
        status: { type: string }
        activities:
          type: array
          items: { type: string }
        discordLink: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        segment: { type: string }
    LoreEntry:
      type: object
      required: [id, title, summary, sortOrder]
      properties:
        id: { type: integer }
        title: { type: string }
        summary: { type: string }
        sortOrder: { type: integer }
        imageUrl: { type: ["string", "null"] }
    LoreInput:
      type: object
      required: [title, summary]
      properties:
        title: { type: string }
        summary: { type: string }
        sortOrder: { type: integer }
        imageUrl: { type: ["string", "null"] }
    LoreUpdate:
      type: object
      properties:
        title: { type: string }
        summary: { type: string }
        sortOrder: { type: integer }
        imageUrl: { type: ["string", "null"] }
...
    GalleryItem:
      type: object
      required: [id, imageUrl, description, sortOrder]
      properties:
        id: { type: integer }
        imageUrl: { type: string }
        description: { type: string }
        sortOrder: { type: integer }
    GalleryItemInput:
      type: object
      required: [imageUrl, description]
      properties:
        imageUrl: { type: string }
        description: { type: string }
        sortOrder: { type: integer }
    GalleryItemUpdate:
      type: object
      properties:
        imageUrl: { type: string }
        description: { type: string }
        sortOrder: { type: integer }
...
    AdminLoginInput:
      type: object
      required: [username]
      properties:
        username: { type: string }
        password: { type: string }
...