# lib/api-zod/src/generated/types/galleryItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItemInput {
  imageUrl: string;
  description: string;
  sortOrder?: number;
}