# lib/api-zod/src/generated/types/illegalOrgInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IllegalOrgInput {
  name: string;
  status: string;
  activities: string[];
  /** @nullable */
  discordLink?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  segment?: string;
}