# lib/api-zod/src/generated/types/loreUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoreUpdate {
  title?: string;
  summary?: string;
  sortOrder?: number;
  /** @nullable */
  imageUrl?: string | null;
}